moved unit & scale tests into separate directories
