+++ test/scale/operations/xor.function.scale.test.ts
@@ -1,5 +1,5 @@
 import { afterAll, describe, expect, it, jest } from '@jest/globals';
-import { xor } from '../../src';
+import { xor } from '../../../src';
 import {
 	coupleDisjoint,
 	coupleEquivalent,
@@ -14,11 +14,10 @@ import {
 	someDisjoint,
 	someEquivalent,
 	times,
-} from '../util/scale/scale-testing.constants';
-import { Timer } from '../util/scale/timer.model';
+} from '../../util/scale/scale-testing.constants';
+import { Timer } from '../../util/scale/timer.model';
 
-export function xorScaleTests(): void {
-
+describe('xor @ scale', () => {
 	describe('xor ⋅ large sets', () => {
 		it('xor(of1):'.padEnd(padding), () => {
 			const result = Timer.time('xor', () => xor(multiplesOf1));
@@ -105,4 +104,4 @@ export function xorScaleTests(): void {
 	});
 
 	afterAll(() => Timer.log('xor'));
-}
+});
