+++ test/scale/scale.test.ts
@@ -1,0 +1,14 @@
+import './operations/difference.function.scale.test';
+import './operations/intersection.function.scale.test';
+import './operations/union.function.scale.test';
+import './operations/xor.function.scale.test';
+
+import './comparisons/disjoint.function.scale.test';
+import './comparisons/equivalence.function.scale.test';
+import './comparisons/pairwise-disjoint.function.scale.test';
+import './comparisons/proper-subset.function.scale.test';
+import './comparisons/proper-superset.function.scale.test';
+import './comparisons/subset.function.scale.test';
+import './comparisons/superset.function.scale.test';
+
+import './ordering/sort.function.scale.test';
