+++ test/unit/comparisons/superset.function.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from '@jest/globals';
-import { superset } from '../../src';
-import { TestSets } from '../util/test-sets/test-sets.model';
-import { testSuite } from '../util/test-suite.function';
+import { superset } from '../../../src';
+import { TestSets } from '../../util/test-sets/test-sets.model';
+import { testSuite } from '../../util/test-suite.function';
 
 describe('superset', () => {
 	testSuite('superset', supersetTests);
