+++ test/unit/operations/difference.function.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from '@jest/globals';
-import { difference, equivalence } from '../../src';
-import { TestSets } from '../util/test-sets/test-sets.model';
-import { testSuite } from '../util/test-suite.function';
+import { difference, equivalence } from '../../../src';
+import { TestSets } from '../../util/test-sets/test-sets.model';
+import { testSuite } from '../../util/test-suite.function';
 
 describe('difference', () => {
 	testSuite('difference', differenceTests);
